[CCA-1] Imports the basic framework structure with adapted Selenium examples.
Being new to Selenium, I took on the Getting Started Selenium examples, but I already had in mind
that I wanted to make sure that different browsers could run the Selenium tests. Therefore, I added
the challenge of wrapping the WebDrivers into factory classes, which allowed for configuration via
Properties files.

I messed around with differing versions of the Selenium dependencies, as I kept observing the
HtmlUnitDriver completing the tests, but then hanging around for a slice of time much longer than I
expected, especially in the face of the graphical FirefoxDriver (which was an adventure getting to
work properly). So there has to be more improvements to be made.

I still have to create the README, and also add some Maven goals to compile the JAR file that is
referenced in the usage instructions. Oh, and the actual assessment itself. The examples included
in this commit are just to demonstrate that the mini-framework I put together works. There's much
more to do!