[CCA-2] Adds first assessment Selenium test.
The first test verifies that when the page is scrolled downwards, the navigation bar background loses the
transparent-nav class style.

Tested with the FirefoxDriver, although there are some JavaScript errors in the log. The other two WebDrivers
fail, due to the lack of JavaScript being enabled in the HtmlUnitDriver by default behavior, and in the
HtmlUnitWithJsDriver due to additional Js errors (most likely related to how Rhino works).

This test could be expanded/duplicated to verify that when the user/agent scrolls back up to the top, the
transparent-nav class style is added back to the header element.

As an asside, the two example Selenium tests adapted from Getting Started with Selenium, as they are no longer
the focus of this project.