[CCA-3] Adds the second Selenium test case.
This test case checks for equality between the two Advertiser Disclosures on the landing page, the first
behind a label/link that requires the user to click to see, and the second in the footer of the page.

Bonus Achieved! They aren't the same. Which one is the more correct? I believe they have the same... ah..
(non-legalese) intent (?) behind them, so I don't think it's too big a deal, but still, would be nice to
have the same disclosure.

At least there is a nice link with sizable text not buried in the footer to inform the user about the
Disclosure, which is really nice!