[CCA-5] Adds the fourth Selenium test, a little more involved.
This tests the user wanting to see cards offering to save on gas purchases.

It definitely helps to debug which element is the reference to moving forward in the flow, when
the logic seems right, but the test fails. Also threw in some/more Java Stream API, which makes
it look subjectively better.